+++ fiddles/angular2-cli/fiddle-0031-NgxBootstrapBasics/src/app/components/modal/modal.component.ts
@@ -1,0 +1,20 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-modal',
+  templateUrl: './modal.component.html',
+  styleUrls: ['./modal.component.css']
+})
+export class ModalComponent {
+
+  isModal: boolean;
+
+  onShowModal() {
+    this.isModal = true;
+  }
+
+  onPopUpEvents(event: boolean) {
+    this.isModal = event;
+  }
+
+}