+++ fiddles/angular2-cli/fiddle-0031-NgxBootstrapBasics/src/app/components/panel-menu/dropdown/dropdown.options.ts
@@ -1,0 +1,21 @@
+export interface Dropdown {
+  /**
+   * Unique identifier.
+   */
+  id?: string;
+
+  /**
+   * The text which will be displayed.
+   */
+  text?: string;
+
+  /**
+   * Data being passed.
+   */
+  data?: any;
+
+  /**
+   * pseudo class added to the menu link.
+   */
+  cssClass?: string;
+}