+++ fiddles/angular2-cli/fiddle-0031-NgxBootstrapBasics/src/app/components/popover/popover.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { PopoverComponent } from './popover.component';
+import { PopoverModule as BsPopoverModule } from 'ngx-bootstrap/popover';
+
+
+@NgModule({
+  declarations: [PopoverComponent],
+  imports: [
+    CommonModule,
+    BsPopoverModule.forRoot()
+  ],
+  exports: [PopoverComponent],
+  entryComponents: [PopoverComponent]
+})
+export class PopoverModule { }