+++ fiddles/angular2-cli/fiddle-0032-ImageLazyLoad/src/app/components/components.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { PhotoViewerComponent } from './photo-viewer/photo-viewer.component';
+
+
+
+@NgModule({
+  declarations: [PhotoViewerComponent],
+  imports: [
+    CommonModule
+  ],
+  exports: [
+    PhotoViewerComponent
+  ]
+})
+export class ComponentsModule { }