+++ CHANGELOG.md
@@ -2204,3 +2204,17 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **Angular-Cli**
 
     1. Added [fiddles/angular2-cli/fiddle-0030-EntAgGridChartRangeSelect](fiddles/angular2-cli/fiddle-0030-EntAgGridChartRangeSelect)
+
+
+### 326_AngularCli_31 ~ [Issue 326](https://github.com/bradyhouse/house/issues/326)
+
+* **Angular-Cli**
+
+    1. Added [fiddles/angular2-cli/fiddle-0031-NgxBootstrapBasics](fiddles/angular2-cli/fiddle-0031-NgxBootstrapBasics)
+
+
+### 328_AngularCli_32 ~ [Issue 328](https://github.com/bradyhouse/house/issues/328)
+
+* **Angular-Cli**
+
+    1. Added [fiddles/angular2-cli/fiddle-0032-ImageLazyLoad](fiddles/angular2-cli/fiddle-0032-ImageLazyLoad)